+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 * nostr: add `Tags::challenge` method ([Yuki Kishimoto])
 * nostr: add `RelayUrl::is_local_addr` ([Yuki Kishimoto])
 * nostr: add `TagKind::k` constructor ([Yuki Kishimoto]) 
+* nostr: impl `IntoIterator` for `Tag` ([Yuki Kishimoto])
 * database: impl PartialEq and Eq for `Events` ([Yuki Kishimoto])
 * database: add `SaveEventStatus` enum ([Yuki Kishimoto])
 * pool: add `ReceiverStream` ([Yuki Kishimoto])